RicoBullishTestBot/requirements.txt

PyYAML>=6.0
rich>=12.5.1
TgCrypto>=1.2.3
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.20.0
google-cloud-core>=2.4.0
pyrogram>=2.0.106


APScheduler~=3.10.4

Telethon>=1.36.0
python-dotenv>=1.0.1
protobuf>=5.27.3
pytz>=2024.1
pillow>=10.4.0
tqdm>=4.66.4
